web module commit

+++ op-user/op-user-domain/src/main/java/com/kkk/op/user/domain/service/impl/AccountServiceImpl.java
@@ -4,32 +4,47 @@ import com.kkk.op.support.type.LongId;
 import com.kkk.op.user.domain.entity.Account;
 import com.kkk.op.user.domain.service.AccountService;
 import com.kkk.op.user.repository.AccountRepository;
-import javax.validation.constraints.NotNull;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 /**
  *
  * @author KaiKoo
  */
+@Service
 public class AccountServiceImpl implements AccountService {
 
     private final AccountRepository accountRepository;
 
-    public AccountServiceImpl(AccountRepository accountRepository) {
+    public AccountServiceImpl(@Autowired AccountRepository accountRepository) {
         this.accountRepository = accountRepository;
     }
 
     @Override
-    public Account find(@NotNull LongId id) {
+    public Account find(LongId id) {
+        if (id == null) {
+            return null;
+        }
         return accountRepository.find(id);
     }
 
     @Override
-    public void remove(@NotNull Account entity) {
+    public void remove(Account entity) {
+        Account account = this.find(entity.getId());
+        // 逻辑校验
+        if (account == null) {
+            throw new IllegalArgumentException("id不存在");
+        }
+        // 移除
         accountRepository.remove(entity);
     }
 
     @Override
-    public void save(@NotNull Account entity) {
+    public void save(Account entity) {
+        Account account = this.find(entity.getId());
+        if (account == null && entity.getId() != null) {
+            throw new IllegalArgumentException("id不存在");
+        }
         accountRepository.save(entity);
     }
 }
